No more partial class Drawable
Also makes the name "Transform" consistent, rather than sometimes
calling it "Transformation".
