+++ osu.Framework/Graphics/Containers/Container_AutoSize.cs
@@ -5,7 +5,7 @@ using System;
 using osu.Framework.Caching;
 using OpenTK;
 using osu.Framework.Graphics.Primitives;
-using osu.Framework.Graphics.Transformations;
+using osu.Framework.Graphics.Transforms;
 using System.Linq;
 
 namespace osu.Framework.Graphics.Containers
