+++ osu.Framework/Graphics/Transforms/TransformColour.cs
@@ -5,7 +5,7 @@ using osu.Framework.MathUtils;
 using OpenTK.Graphics;
 using osu.Framework.Graphics.Colour;
 
-namespace osu.Framework.Graphics.Transformations
+namespace osu.Framework.Graphics.Transforms
 {
     /// <summary>
     /// Transforms colour value in linear colour space.
