+++ osu.Framework/Graphics/Transforms/TransformGlowAlpha.cs
@@ -3,7 +3,7 @@
 
 using osu.Framework.Graphics.Containers;
 
-namespace osu.Framework.Graphics.Transformations
+namespace osu.Framework.Graphics.Transforms
 {
     public class TransformGlowAlpha : TransformFloat
     {
