+++ osu.Framework/Graphics/Visualisation/LogOverlay.cs
@@ -4,7 +4,7 @@
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Primitives;
 using osu.Framework.Graphics.Sprites;
-using osu.Framework.Graphics.Transformations;
+using osu.Framework.Graphics.Transforms;
 using osu.Framework.Logging;
 using OpenTK;
 using OpenTK.Graphics;
