+++ osu.Framework/MathUtils/Interpolation.cs
@@ -3,7 +3,7 @@
 
 using System;
 using osu.Framework.Graphics.Primitives;
-using osu.Framework.Graphics.Transformations;
+using osu.Framework.Graphics.Transforms;
 using OpenTK;
 using OpenTK.Graphics;
